Allow search for both org_name and location
Not sure exactly what’s going on here, but it looks eager loading the
organization table when both the org_name and location parameters are
present causes an invalid SQL. The solution is to conditionally eager
load the organization table only if both parameters are not present.

There probably is a better solution but this does the trick for now.